fix hydration error & extract playback indicator
