+++ src/components/templates/andNavigation.tsx
@@ -8,27 +8,24 @@ import React, {
 	useRef,
 	useState,
 } from 'react';
-import { FormattedMessage, useIntl } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 
 import Button from '@components/molecules/button';
-import IconButton from '@components/molecules/iconButton';
 import LanguageAlternativesAlert from '@components/molecules/languageAlternativesAlert';
 import Mininav from '@components/molecules/mininav';
 import SearchBar from '@components/molecules/searchBar';
 import Header from '@components/organisms/header';
 import Navigation from '@components/organisms/navigation';
-import { BaseColors } from '@lib/constants';
 import isServerSide from '@lib/isServerSide';
 import { makeDonateRoute } from '@lib/routes';
 import useLanguageRoute from '@lib/useLanguageRoute';
 import { useNavigationItems } from '@lib/useNavigationItems';
 
-import IconListeningAnimated from '../../../public/img/icons/icon-listening-animated.svg';
-import IconListening from '../../../public/img/icons/icon-listening.svg';
 import MoreIcon from '../../../public/img/icons/icon-more.svg';
 
 import styles from './andNavigation.module.scss';
 import { PlaybackContext } from './andPlaybackContext';
+import ButtonPlayback from '@components/molecules/buttonPlayback';
 
 const SUBNAV_HEIGHT = 32;
 const HEADER_TITLE_PADDING_TOP_DIFFERENCE = 8;
@@ -44,9 +41,7 @@ export default function AndNavigation({
 	children: ReactNode;
 }): JSX.Element {
 	const router = useRouter();
-	const intl = useIntl();
 	const {
-		push,
 		query: { q },
 		pathname,
 		asPath,
@@ -140,21 +135,7 @@ export default function AndNavigation({
 						}
 						href={makeDonateRoute(languageRoute)}
 					/>
-					{/* TODO: extract media playback indicator */}
-					{playbackRecording && (
-						<IconButton
-							Icon={
-								playbackContext.paused() ? IconListening : IconListeningAnimated
-							}
-							onClick={() => push(playbackRecording.canonicalPath || '')}
-							color={BaseColors.RED}
-							backgroundColor={BaseColors.CREAM}
-							aria-label={intl.formatMessage({
-								id: 'andNavigation__playing',
-								defaultMessage: 'Playing',
-							})}
-						/>
-					)}
+					{playbackRecording && <ButtonPlayback />}
 				</div>
 				<div className={styles.mobileSubnavWrapper} ref={subnavRef}>
 					<div className={styles.mobileSubnav}>
